QIBC_53BP1 foci_cell cycle
Code for a quantitative image-based cytometry (QIBC) method to quantify the number of 53BP1 foci in individual nuclei. This code is used to create a specific QIBC cell-cycle profile based on DAPI (marker of DNA content) and EdU (marker of replicating cells) intensity in cells. Cell cycle phases are divided into the following: G1. Early S, Mid S, Late S and G2. For this particular experiment, 53BP1 foci (per area) were counted in individual U2OS cells and cell cycle profile was created to study the changes in the number of 53BP1 foci in each of the cell cycle phases. In this code, “TIRR KO” signifies one of the experimental conditions (in which our target of interest, TIRR, was knocked out using CRISPR). The same code was used for the wild-type (WT) condition, in which case “TIRR KO” can be replaced in the code for “WT”. For ggpolt x-axis shows total DAPI intensity, y-axis shows mean EdU intensity, and ‘color’ signifies the EdU mean intensity. 'Color' can be changed to say '53BP1 foci' to get the plot outlining 53BP1 foci number in different cell cycle phases.